--- src/stream/src/executor/exchange.rs
@@ -1,1 +1,0 @@
-pub mod input;
