+++ src/stream/src/executor/mod.rs
@@ -49,7 +49,7 @@ mod barrier_align;
 mod batch_query;
 mod chain;
 mod debug;
-pub mod dispatch;
+mod dispatch;
 mod dynamic_filter;
 mod error;
 pub mod exchange;
@@ -63,13 +63,13 @@ mod local_simple_agg;
 mod lookup;
 mod lookup_union;
 mod managed_state;
-pub mod merge;
+mod merge;
 pub mod monitor;
 mod mview;
 mod project;
 mod project_set;
 mod rearranged_chain;
-pub mod receiver;
+mod receiver;
 mod simple;
 mod sink;
 mod source;
@@ -87,7 +87,7 @@ pub use actor::{Actor, ActorContext, ActorContextRef, OperatorInfo, OperatorInfo
 pub use batch_query::BatchQueryExecutor;
 pub use chain::ChainExecutor;
 pub use debug::DebugExecutor;
-pub use dispatch::DispatchExecutor;
+pub use dispatch::{DispatchExecutor, DispatcherImpl};
 pub use dynamic_filter::DynamicFilterExecutor;
 pub use expand::ExpandExecutor;
 pub use filter::FilterExecutor;
@@ -103,6 +103,7 @@ pub use mview::*;
 pub use project::ProjectExecutor;
 pub use project_set::*;
 pub use rearranged_chain::RearrangedChainExecutor;
+pub use receiver::ReceiverExecutor;
 use risingwave_pb::source::{ConnectorSplit, ConnectorSplits};
 use simple::{SimpleExecutor, SimpleExecutorWrapper};
 pub use sink::SinkExecutor;
